+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
 .vscode
-test-recipes.html
 config.py
 __pycache__
+test-recipes.html
