+++ Dockerfile
@@ -1,0 +1,11 @@
+FROM python:3.11
+
+WORKDIR /app
+
+COPY requirements.txt requirements.txt
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+CMD ["python", "app.py"]
