+++ get_recipes.py
@@ -1,7 +1,6 @@
 import requests
 from bs4 import BeautifulSoup
 from urllib.parse import urljoin, quote
-from jinja2 import Template
 
 RAKUTEN_BASE_URL = "https://recipe.rakuten.co.jp" # 楽天レシピのURL
 
@@ -37,55 +36,3 @@ def fetch_recipes(ingredients): # レシピの取得
             print(f"Invalid recipe: {new_recipe}")
     
     return recipes
-
-ingredients = ['卵', '牛乳'] # ここに材料を記述
-recipes = fetch_recipes(ingredients)
-
-# 取得したものをHTMLファイルに出力
-template = Template('''
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Recipes</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 20px;
-        }
-        .recipe {
-            border: 1px solid #ddd;
-            margin-bottom: 20px;
-            padding: 10px;
-            border-radius: 5px;
-        }
-        .recipe img {
-            max-width: 100px;
-            height: auto;
-        }
-        .recipe h2 {
-            margin: 0;
-            font-size: 1.5em;
-        }
-    </style>
-</head>
-<body>
-    <h1>Recipes</h1>
-    {% for recipe in recipes %}
-    <div class="recipe">
-        <h2>{{ recipe.title }}</h2>
-        <img src="{{ recipe.image }}" alt="{{ recipe.title }}">
-        <p><a href="{{ recipe.url }}">View Recipe</a></p>
-    </div>
-    {% endfor %}
-</body>
-</html>
-''')
-
-html_output = template.render(recipes=recipes)
-
-with open('test/test-recipes.html', 'w', encoding='utf-8') as f:
-    f.write(html_output)
-
-print("Success!")
