+++ requirements.txt
@@ -2,3 +2,5 @@ requests
 beautifulsoup4
 jinja2
 openai
+flask
+flask-cors
